refactor: テストを tests/ 配下に集約 (Phase 2/4, #109)
- src/__tests__/ の unit/component テストを tests/unit/ と tests/component/ に移動
- src/test/setup.ts を tests/setup/vitest.ts に移動
- plugins/__tests__/ を tests/unit/tooling/ に移動
- vite.config.ts: setupFiles パスと include パターンを更新
- tsconfig.app.json: tests を含まない src include のため exclude を削除
- 各テストファイルの import/vi.mock パスを ../../src/ 形式に書き換え
- 空ディレクトリ (src/__tests__, src/test, plugins/__tests__) を削除
