+++ tsconfig.app.json
@@ -24,6 +24,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"],
-  "exclude": ["src/__tests__", "src/test"]
+  "include": ["src"]
 }
